swapped out tqdm for log_progress functions across a number of modules

+++ thickshake/storage/writer.py
@@ -8,20 +8,20 @@
 import csv
 import json
 import logging
+import os
 import time
 
 ##########################################################
 # Third Party Imports
 
 import h5py
+from tqdm import tqdm
 
 ##########################################################
 # Local Imports
 
-from thickshake.helpers import (
-    open_file, json_serial, log_progress,
-    setup_logging, setup_warnings, get_file_type,
-)
+from thickshake.storage.store import Store
+from thickshake.helpers import open_file, json_serial, get_file_type, FileType
 
 ##########################################################
 # Typing Configuration
@@ -35,11 +35,6 @@ JSONType = Union[Dict[str, Any], List[Any]]
 ##########################################################
 # Environmental Variables
 
-class FileType:
-    JSON = ".json"
-    HDF5 = ".hdf5"
-    XML = ".xml"
-    CSV = ".csv"
 
 ##########################################################
 # Logging Configuration
@@ -56,7 +51,7 @@ def write_csv(records: List[Any], output_file: FilePath, **kwargs: Any) -> None:
         outcsv = csv.writer(outfile)
         header = records[0].keys()
         outcsv.writerow(header)
-        for record in records:
+        for record in tqdm(records, desc="Writing Records"):
             record_list = record.values()
             outcsv.writerow(record_list)
 
@@ -64,27 +59,26 @@ def write_csv(records: List[Any], output_file: FilePath, **kwargs: Any) -> None:
 def write_json(records: JSONType, output_file: FilePath, **kwargs: Any) -> None:
     if not records: return None
     with open_file(output_file, 'w+', encoding='utf-8') as outfile:
-        json.dump(records, outfile, indent=2, default=json_serial)
+        outfile.write("[")
+        length = len(records)
+        for i, record in enumerate(tqdm(records, desc="Writing Records")):
+            json_obj = json.dump(record, outfile, indent=2, default=json_serial)
+            if i < length -1: outfile.write(",")
+        outfile.write("]")
 
 
-def write_hdf5(records: List[Any], output_file: FilePath, logging_flag: bool = True, **kwargs: Any) -> None:
+def write_hdf5(records: List[Any], output_file: FilePath, **kwargs: Any) -> None:
     with h5py.File(output_file, "a") as f:
         total = len(records)
         start_time = time.time()
-        for i, record in enumerate(records):
-            key = record["image_id"]
+        for record in tqdm(records, desc="Writing Records"):
+            key = str(record["uuid"])
             grp = f.require_group(key)
             index = str(len(grp.keys()) + 1)
             dt = h5py.special_dtype(vlen=str)
             f.attrs.create("columns", list(record.keys()),dtype=dt)
             record_serial = json.dumps(record, default=json_serial)
             grp.require_dataset(index, data=record_serial, shape=(1,), dtype=dt)
-            if logging_flag: log_progress(i+1, total, start_time, interval=100)
-
-
-#TODO
-def write_xml(records: List[Any], output_file: FilePath, **kwargs: Any) -> None:
-    pass
 
 
 def write_log(records: List[Any], **kwargs: Any) -> None:
@@ -98,24 +92,24 @@ def write_to_store(records: List[Any], store: Store, **kwargs: Any) -> None:
     pass
 
 
-def write_flat_file(records: List[Any], output_file: FilePath, **kwargs: Any) -> None:
+def write_flat_file(records: List[Any], output_file: FilePath, force: bool=False, **kwargs: Any) -> None:
+    if not force and os.path.exists(output_file): raise IOError
     file_type = get_file_type(output_file) #DONE
     if file_type == FileType.JSON:
         write_json(records, output_file, **kwargs) #DONE
     elif file_type == FileType.HDF5:
         write_hdf5(records, output_file, **kwargs) #DONE
-    elif file_type == FileType.XML:
-        write_xml(records, output_file, **kwargs) #TODO
     elif file_type == FileType.CSV:
         write_csv(records, output_file, **kwargs) #DONE
     else: raise NotImplementedError
 
 
-def export_flat_file(output_file: FilePath, **kwargs: Any) -> None:
+def export_flat_file(output_file: FilePath, force: bool=False, **kwargs: Any) -> None:
+    if not force and os.path.exists(output_file): raise IOError
     from thickshake.storage.database import Database
     database = Database()
-    records = database.dump()
-    write_flat_file(records, output_file, **kwargs)
+    records = database.dump(force=force, **kwargs)
+    write_flat_file(records, output_file, force=force, **kwargs)
 
 
 def export_to_store(**kwargs: Any) -> None:
